Seperate components from derive module
